reorg dir, move handlers out
