+++ frontend/lib/src/components/widgets/ChatInput/ChatInput.tsx
@@ -23,21 +23,14 @@ import React, {
   useState,
 } from "react"
 
-import axios from "axios"
 import { useTheme } from "@emotion/react"
 import { Send } from "@emotion-icons/material-rounded"
 import { Textarea as UITextArea } from "baseui/textarea"
-import {
-  ErrorCode as FileErrorCode,
-  FileRejection,
-  useDropzone,
-} from "react-dropzone"
-import zip from "lodash/zip"
+import { useDropzone } from "react-dropzone"
 
 import {
   ChatInput as ChatInputProto,
   FileUploaderState as FileUploaderStateProto,
-  FileURLs as FileURLsProto,
   IChatInputValue,
   IFileURLs,
   UploadedFileInfo as UploadedFileInfoProto,
@@ -48,7 +41,7 @@ import {
   chatInputAcceptFileProtoValueToEnum,
   isNullOrUndefined,
 } from "~lib/util/utils"
-import { WidgetInfo, WidgetStateManager } from "~lib/WidgetStateManager"
+import { WidgetStateManager } from "~lib/WidgetStateManager"
 import Icon from "~lib/components/shared/Icon"
 import InputInstructions from "~lib/components/shared/InputInstructions/InputInstructions"
 import { isEnterKeyPressed } from "~lib/util/inputUtils"
@@ -65,8 +58,10 @@ import {
   StyledSendIconButton,
   StyledSendIconButtonContainer,
 } from "./styled-components"
-import ChatUploadedFiles from "./UploadedFile/ChatUploadedFiles"
-import FileUploadArea from "./FileUploadArea/FileUploadArea"
+import ChatUploadedFiles from "./fileUpload/ChatUploadedFiles"
+import FileUploadArea from "./fileUpload/FileUploadArea"
+import { createUploadFileHandler } from "./fileUpload/createFileUploadHandler"
+import { createDropHandler } from "./fileUpload/createDropHandler"
 
 export interface Props {
   disabled: boolean
@@ -76,155 +71,6 @@ export interface Props {
   uploadClient: FileUploadClient
   fragmentId?: string
 }
-
-interface CreateDropHandlerParams {
-  acceptMultipleFiles: boolean
-  uploadClient: FileUploadClient
-  uploadFile: (fileURLs: FileURLsProto, file: File) => void
-  addFiles: (files: UploadFileInfo[]) => void
-  getNextLocalFileId: () => number
-  deleteExistingFiles: () => void
-  onUploadComplete: () => void
-}
-
-const createDropHandler =
-  ({
-    acceptMultipleFiles,
-    uploadClient,
-    uploadFile,
-    addFiles,
-    getNextLocalFileId,
-    deleteExistingFiles,
-    onUploadComplete,
-  }: CreateDropHandlerParams) =>
-  (acceptedFiles: File[], rejectedFiles: FileRejection[]): void => {
-    // If only single file upload is allowed but multiple were dropped/selected,
-    // all files will be rejected by default. In this case, we take the first
-    // valid file into acceptedFiles, and reject the rest.
-    if (
-      !acceptMultipleFiles &&
-      acceptedFiles.length === 0 &&
-      rejectedFiles.length > 1
-    ) {
-      const firstFileIndex = rejectedFiles.findIndex(
-        file => file.errors?.[0].code === FileErrorCode.TooManyFiles
-      )
-
-      if (firstFileIndex >= 0) {
-        acceptedFiles.push(rejectedFiles[firstFileIndex].file)
-        rejectedFiles.splice(firstFileIndex, 1)
-      }
-    }
-
-    if (!acceptMultipleFiles && acceptedFiles.length > 0) {
-      deleteExistingFiles()
-    }
-
-    uploadClient
-      .fetchFileURLs(acceptedFiles)
-      .then((fileURLsArray: IFileURLs[]) => {
-        zip(fileURLsArray, acceptedFiles).forEach(
-          ([fileURLs, acceptedFile]) => {
-            uploadFile(fileURLs as FileURLsProto, acceptedFile as File)
-          }
-        )
-      })
-      .catch((errorMessage: string) => {
-        addFiles(
-          acceptedFiles.map(f => {
-            return new UploadFileInfo(f.name, f.size, getNextLocalFileId(), {
-              type: "error",
-              errorMessage,
-            })
-          })
-        )
-      })
-
-    // Create an UploadFileInfo for each of our rejected files, and add them to
-    // our state.
-    if (rejectedFiles.length > 0) {
-      const rejectedInfos = rejectedFiles.map(
-        rejected =>
-          new UploadFileInfo(
-            rejected.file.name,
-            rejected.file.size,
-            getNextLocalFileId(),
-            {
-              type: "error",
-              errorMessage: rejected.errors
-                .map(err => err.message)
-                .filter(err => !!err)
-                .join(", "),
-            }
-          )
-      )
-      addFiles(rejectedInfos)
-    }
-
-    onUploadComplete()
-  }
-
-interface CreateUploadFileParams {
-  getNextLocalFileId: () => number
-  addFiles: (files: UploadFileInfo[]) => void
-  updateFile: (id: number, fileInfo: UploadFileInfo) => void
-  uploadClient: FileUploadClient
-  element: WidgetInfo
-  onUploadProgress: (e: ProgressEvent, id: number) => void
-  onUploadComplete: (id: number, fileURLs: IFileURLs) => void
-}
-
-const createUploadFileHandler =
-  ({
-    getNextLocalFileId,
-    addFiles,
-    updateFile,
-    uploadClient,
-    element,
-    onUploadProgress,
-    onUploadComplete,
-  }: CreateUploadFileParams) =>
-  (fileURLs: IFileURLs, file: File): void => {
-    // Create an UploadFileInfo for this file and add it to our state.
-    const cancelToken = axios.CancelToken.source()
-    const uploadingFileInfo = new UploadFileInfo(
-      file.name,
-      file.size,
-      getNextLocalFileId(),
-      {
-        type: "uploading",
-        cancelToken,
-        progress: 1,
-      }
-    )
-    addFiles([uploadingFileInfo])
-
-    uploadClient
-      .uploadFile(
-        {
-          formId: "", // TODO[kajarnec] fix this probably with uploadFile refactoring
-          ...element,
-        },
-        fileURLs.uploadUrl as string,
-        file,
-        e => onUploadProgress(e, uploadingFileInfo.id),
-        cancelToken.token
-      )
-      .then(() => onUploadComplete(uploadingFileInfo.id, fileURLs))
-      .catch(err => {
-        // If this was a cancel error, we don't show the user an error -
-        // the cancellation was in response to an action they took.
-        if (!axios.isCancel(err)) {
-          updateFile(
-            uploadingFileInfo.id,
-            uploadingFileInfo.setStatus({
-              type: "error",
-              errorMessage: err ? err.toString() : "Unknown error",
-            })
-          )
-        }
-      })
-  }
 
 // We want to show easily that there's scrolling so we deliberately choose
 // a half size.
@@ -335,12 +181,12 @@ function ChatInput({
     uploadFile: createUploadFileHandler({
       getNextLocalFileId,
       addFiles,
-      updateFile: (id, fileInfo) => {
+      updateFile: (id: number, fileInfo: UploadFileInfo) => {
         setFiles(files => updateFile(id, fileInfo, files))
       },
       uploadClient,
       element,
-      onUploadProgress: (e, fileId) => {
+      onUploadProgress: (e: ProgressEvent, fileId: number) => {
         setFiles(files => {
           const file = getFile(fileId, files)
           if (isNullOrUndefined(file) || file.status.type !== "uploading") {
@@ -363,7 +209,7 @@ function ChatInput({
           )
         })
       },
-      onUploadComplete: (id, fileUrls) => {
+      onUploadComplete: (id: number, fileUrls: IFileURLs) => {
         setFiles(files => {
           const curFile = getFile(id, files)
           if (
