+++ frontend/lib/src/components/widgets/ChatInput/fileUpload/styled-components.ts
